Move Circus types into @jest/types
