+++ packages/jest-types/src/index.ts
@@ -5,7 +5,8 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import * as Circus from './Circus';
 import * as Config from './Config';
 import * as Global from './Global';
 
-export {Config, Global};
+export {Circus, Config, Global};
